Hacked out the utilities from stats/utils.
